backend/requirements.txt
Django==4.0.5
djangorestframework==3.13.1
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
gunicorn==20.1.0